collection_module/collection_ui/src/main/java/in/okcredit/collection_ui/usecase/EnablePaymentAddressImpl.kt
package `in`.okcredit.collection_ui.usecase

import `in`.okcredit.collection.contract.CollectionRepository
import `in`.okcredit.collection.contract.EnablePaymentAddress
import `in`.okcredit.merchant.contract.GetActiveBusinessId
import dagger.Lazy
import io.reactivex.Completable
import javax.inject.Inject

class EnablePaymentAddressImpl @Inject constructor(
    private val collectionRepository: CollectionRepository,
    private val getActiveBusinessId: Lazy<GetActiveBusinessId>,
) : EnablePaymentAddress {

    override fun execute(): Completable {
        return getActiveBusinessId.get().execute().flatMapCompletable { businessId ->
            collectionRepository.enableCustomerPayment(businessId)
        }
    }
}
